+++ index.js
@@ -1,5 +1,6 @@
-var Factory, strgMethods;
+var Factory, strgMethods, mongoose;
 strgMethods = require('strg_methods');
+mongoose = require("mongoose");
 
 /**
 @class Factory
@@ -22,7 +23,11 @@ strgMethods = require('strg_methods');
 @default 0
 */
 function Factory(model, factory){
-  this.model = model;
+  if (typeof model == "string"){
+    this.model = mongoose.model(model);
+  } else {
+    this.model = model;
+  }
   this.factory = factory || model;
   this.sequenc = 0;
 }
@@ -58,6 +63,26 @@ Factory.prototype.build = function (options, callback){
 
 /**
 
+returns keys of an nested object
+
+@method _getKeys
+@private
+@param obj {Object} object nested or not
+@return {Array} returns array of obj and nested obj keys
+*/
+Factory.prototype._getKeys = function (obj) {
+  var arr = [];
+  for (var attrname in obj) {
+    arr.push(attrname);
+    if (typeof obj[attrname] === "object" && attrname != "_id"){
+      arr = arr.concat(this._getKeys(obj[attrname]));
+    }
+  }
+  return arr;
+};
+
+/**
+
 checks if new mongoose.model(object) got a wrong field value
 
 @method _compareObjSync
@@ -70,8 +95,9 @@ Factory.prototype._compareObjSync = function (mObj, obj) {
   var defaults, arr, mArr;
   if(this.model === this.factory){ return null; }
   obj._id = "";
-  mArr = Object.keys(mObj.toObject()).sort();
-  arr = Object.keys(obj).sort();
+  mObj._id = "";
+  mArr = this._getKeys(mObj.toObject()).sort();
+  arr = this._getKeys(obj).sort();
   //check if default by adding keys from mObj to obj
   defaults = mArr.filter(function(val) {
     return arr.indexOf(val) == -1;
@@ -99,10 +125,18 @@ Factory.prototype._mergeObjsSync = function (obj1, obj2) {
   var attrname, obj3;
   obj3 = {};
   for (attrname in obj1) {
-    obj3[attrname] = obj1[attrname];
+    if (typeof obj1[attrname] === "object"){
+      obj3[attrname] = this._mergeObjsSync(obj1[attrname], obj2[attrname]);
+    } else {
+      obj3[attrname] = obj1[attrname];
+    }
   }
   for (attrname in obj2) {
-    obj3[attrname] = obj2[attrname];
+    if (typeof obj2[attrname] === "object"){
+      obj3[attrname] = this._mergeObjsSync(obj1[attrname], obj2[attrname]);
+    } else {
+      obj3[attrname] = obj2[attrname];
+    }
   }
   return obj3;
 };
@@ -188,9 +222,9 @@ set new doc
 */
 Factory.prototype._newDoc = function (factory, doc){
   if(this.model === this.factory){
-    return this.stringMethods(this._mergeObjsSync(factory, doc));
+    return this._stringMethods(this._mergeObjsSync(factory, doc));
   } else {
-    return new this.model(this.stringMethods(this._mergeObjsSync(factory, doc)));
+    return new this.model(this._stringMethods(this._mergeObjsSync(factory, doc)));
   }
 };
 
@@ -198,17 +232,21 @@ Factory.prototype._newDoc = function (factory, doc){
 
 applies strMethods.all() to each value of object
 
-@method stringMethods
+@method _stringMethods
 @uses strgMethods
 @private
 @param doc {Object} takes a document
 @return {Object} object with applied strgMethods
 */
-Factory.prototype.stringMethods = function (doc){
+Factory.prototype._stringMethods = function (doc){
   for (var key in doc) {
     if (doc.hasOwnProperty(key)) {
-      Strg = new strgMethods(doc[key], this.sequenc);
-      doc[key] = Strg.all();
+      if (typeof doc[key] === "object"){
+        doc[key] = this._stringMethods(doc[key]);
+      } else {
+        Strg = new strgMethods(doc[key], this.sequenc);
+        doc[key] = Strg.all();
+      }
     }
   }
   return doc;
@@ -249,45 +287,6 @@ Factory.prototype.create = function (options, callback) {
       }
     });
   });
-};
-
-/**
-
-instead of factory.model.find
-see http://mongoosejs.com/docs/api.html#model_Model.find
-
-@method find
-@uses mongoose.model.find
-*/
-Factory.prototype.find =  function (args) {
-  if(this.model === this.factory){ throw new Error("you cant use a mongoose method on a plain factory object"); }
-  this.model.find.call(arguments);
-};
-
-/**
-
-instead of factory.model.count
-see http://mongoosejs.com/docs/api.html#model_Model.count
-
-@method count
-@uses mongoose.model.count
-*/
-Factory.prototype.count =  function (args) {
-  if(this.model === this.factory){ throw new Error("you cant use a mongoose method on a plain factory object"); }
-  this.model.count.call(arguments);
-};
-
-/** 
-
-instead of factory.model.findOne
-see http://mongoosejs.com/docs/api.html#model_Model.findOne
-
-@method findOne
-@uses mongoose.model.findOne
-*/
-Factory.prototype.findOne = function (args) {
-  if(this.model === this.factory){ throw new Error("you cant use a mongoose method on a plain factory object"); }
-  this.model.findOne.call(arguments);
 };
 
 /**